added card background to each component
