--- frontend/src/components/CardContainer.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function CardContainer() {
-  return <div className="Card"></div>;
-}
-
-export default CardContainer;
