+++ frontend/src/components/Debt.js
@@ -9,16 +9,20 @@ function Debt({ financeStep, setFinanceStep, debt, setDebt }) {
   };
   return (
     <>
-      <h1>
-        Do you have any short-term high-interest debt? ie. credit cards or
-        personal loans
-      </h1>
-      <button type="button" onClick={yesHandleClick}>
-        Yes
-      </button>
-      <button type="button" onClick={noHandleClick}>
-        No
-      </button>
+      <div className="question-container">
+        <h1>
+          Do you have any short-term high-interest debt? ie. credit cards or
+          personal loans
+        </h1>
+      </div>
+      <div className="btn-container">
+        <button type="button" onClick={yesHandleClick}>
+          Yes
+        </button>
+        <button type="button" onClick={noHandleClick}>
+          No
+        </button>
+      </div>
     </>
   );
 }
