+++ frontend/src/components/NoSavings.js
@@ -7,11 +7,15 @@ function NoSavings({ financeStep, setFinanceStep }) {
 
   return (
     <>
-      <h1>Save up at least 6 months of exspenses before moving on</h1>
+      <div className="question-container">
+        <h1>Save up at least 6 months of exspenses before moving on</h1>
+      </div>
       <h2>maybe insert money tracker here?</h2>
-      <button type="button" onClick={handleClick}>
-        I've saved up enough for 6 months of expenses incase of an emergency!
-      </button>
+      <div className="question-container">
+        <button type="button" onClick={handleClick}>
+          I've saved up enough for 6 months of expenses incase of an emergency!
+        </button>
+      </div>
     </>
   );
 }
