+++ frontend/src/components/OptionalInvestments.js
@@ -3,10 +3,12 @@ import React from "react";
 function OptionalInvestments() {
   return (
     <>
-      <h1>
-        optional investments - Individual Brokerage Acc - RIETS - Real Estate -
-        Bonds - CD's - Crypto -
-      </h1>
+      <div className="question-container">
+        <h1>
+          optional investments - Individual Brokerage Acc - RIETS - Real Estate
+          - Bonds - CD's - Crypto -
+        </h1>
+      </div>
       <a href="https://www.fidelity.com/learning-center/personal-finance/mega-backdoor-roth">
         Theres advanced strategies to investing in 401k this if for those above
         income limits known as MEGA Backdoor ROTH!!!
