+++ frontend/src/components/Savings.js
@@ -9,16 +9,19 @@ function Savings({ financeStep, setFinanceStep, noSavings, setNoSavings }) {
   };
   return (
     <>
-      <h1>
-        Do you have at least 6 months worth of savings to cover expenses? yes or
-        no?
-      </h1>
-      <button type="button" onClick={yesHandleClick}>
-        Yes
-      </button>
-      <button type="button" onClick={noHandleClick}>
-        No
-      </button>
+      <div className="question-container">
+        <h1>
+          Do you have at least 6 months worth of savings to cover expenses?
+        </h1>
+      </div>
+      <div className="btn-container">
+        <button type="button" onClick={yesHandleClick}>
+          Yes
+        </button>
+        <button type="button" onClick={noHandleClick}>
+          No
+        </button>
+      </div>
     </>
   );
 }
