+++ frontend/src/index.css
@@ -1,3 +1,7 @@
+a {
+  margin-bottom: 5%;
+}
+
 body {
   font: 1em/1.618 Inter, monospace;
 
@@ -27,15 +31,16 @@ button {
 }
 
 .card {
-  max-width: 300px;
-  min-height: 200px;
+  min-height: 450px;
+  max-width: 600px;
+  min-width: 600px;
+
   display: flex;
   flex-direction: column;
   justify-content: space-between;
   text-align: center;
 
-  max-width: 500px;
-  height: 300px;
+  height: 400px;
   padding: 35px;
 
   border: 1px solid rgba(255, 255, 255, 0.25);
